Changes so that TPC-E will compile on diassrv1 (sun4v/SunOS/SunCC)
-  The TPC-E implementation was using arrays whose size was variables (e.g., bool flags[i]). Suprisingly, gcc-4.2+ was accepting it and it was not throwing a compilation error. When moved to diassrv1 and SunCC things changed. As a temporary fix all the arrays became array_guard_t and are dynamically allocated at the heap. This is not good for two reasons:
a) We increase by a lot the rate of new/delete calls. Most probably malloc-related system time will jump to the roof, but we need to have correctness before performance.
b) There may be issues with the 2-d or 3-d arrays that the implementation is using. I have put assertions at the places in two transactions (TRADE_LOOKUP and BROKER_VOLUME) that use such arrays.

We should revise the implementation. The simplest solution is seems to be to substitute those arrays with vectors.