fix for the random-number generation in sparc machines
In Sun CC libraries the RAND_MAX number is defined as 2^15-1 (32767) whereas in linux gcc libraries it's 2^31-1 (2147483647). This means that when we try to create input for tm1 getsubdata for example, if we do URand(0,sf*TM1_SUBS_PER_SF) this will always return something smaller than 32767 and for tm1-10 instead of getting uniform input in the range [0,99999], we are getting uniform input in the range [0,32767]. This can be solved by first randomly picking something in the sf-range [0,10), then in the [0,TM1_SUBS_PER_SF) range, and then multiplying them. In fact, this is how things are done when we use spread so probably that's why I haven't noticed this before. However, in TPC-B we have 100000 as TPCB_ACCOUNTS_PER_BRANCH so this can't work there.

Here our fix is basically to generate several numbers and shift them to be able to produce larger numbers.