(1) a workaround for the xcts that delete a tuple during an index scan to reduce the number of crashes, (2) some code cleanup
For (1);
There was a frequent crash happening when a tuple from a table is deleted while performing an index scan on the same table within the same transaction. The workaround is collecting the tuple RIDs, which I can get during the index scan for the tuples to be deleted, and performing the delete operations after the scan.
Right now, I still observe crashes with the same output. However, before it was crashing very frequently that I hardly even get a 30 secs measurement for more than 1 client. Now, it crashes once in a blue moon; like once or twice during a powerrun on diascld19.
For the interested the crash output is;
assertion failure: sp < smlevel_0::page_sz && sp >= maximum
1. error in page.cpp:2832 Assertion failed
        called from:
	0) page.cpp:2832